Discord.Internal.Gateway

Description

Provides a rather raw interface to the websocket events through a real-time Chan

Documentation

type DiscordHandleGateway = (Chan (Either GatewayException Event), Chan GatewaySendable, IORef (Maybe UpdateStatusOpts))

type DiscordHandleCache = (Chan (Either GatewayException Event), MVar (Either (Cache, GatewayException) Cache))

data GatewayException

Constructors

GatewayExceptionCouldNotConnect Text

GatewayExceptionEventParseError Text Text

GatewayExceptionUnexpected GatewayReceivable Text

GatewayExceptionConnection ConnectionException Text

data Cache

Constructors

Cache

Fields

_currentUser :: User

_dmChannels :: Map ChannelId Channel

_guilds :: Map GuildId (Guild, GuildInfo)

_channels :: Map ChannelId Channel

startCacheThread :: Chan Text -> IO (DiscordHandleCache, ThreadId)

startGatewayThread :: Auth -> DiscordHandleCache -> Chan Text -> IO (DiscordHandleGateway, ThreadId)

Create a Chan for websockets. This creates a thread that writes all the received Events to the Chan

data Auth

Authorization token for the Discord API

data Event

Represents possible events sent by discord. Detailed information can be found at https://discord.com/developers/docs/topics/gateway.

data GatewaySendable