Discord.Internal.Gateway.EventLoop

Description

Provides logic code for interacting with the Discord websocket gateway. Realistically, this is probably lower level than most people will need

Documentation

data GatewayException

Constructors

GatewayExceptionCouldNotConnect Text

GatewayExceptionEventParseError Text Text

GatewayExceptionUnexpected GatewayReceivable Text

GatewayExceptionConnection ConnectionException Text

data ConnLoopState

Constructors

ConnStart

ConnClosed

ConnReconnect Auth Text Integer

connect :: (Connection -> IO a) -> IO a

Securely run a connection IO action. Send a close on exception

type DiscordHandleGateway = (Chan (Either GatewayException Event), Chan GatewaySendable, IORef (Maybe UpdateStatusOpts))

connectionLoop :: Auth -> DiscordHandleGateway -> Chan Text -> IO ()

getPayloadTimeout :: Connection -> Int -> Chan Text -> IO (Either ConnectionException GatewayReceivable)

getPayload :: Connection -> Chan Text -> IO (Either ConnectionException GatewayReceivable)

heartbeat :: Chan GatewaySendable -> Int -> IORef Integer -> IO ()

data ConnectionData

What we need to start an event stream

Constructors

ConnData

Fields

connection :: Connection

connSessionID :: Text

connAuth :: Auth

connChan :: Chan (Either GatewayException Event)

startEventStream :: ConnectionData -> Int -> Integer -> Chan GatewaySendable -> IORef (Maybe UpdateStatusOpts) -> Chan Text -> IO ConnLoopState

eventStream :: ConnectionData -> IORef Integer -> Int -> Chan GatewaySendable -> IORef Bool -> Chan Text -> IO ConnLoopState

data Sendables

Constructors

Sendables

Fields

sendchan :: Chan GatewaySendable

Things the user wants to send. Doesn't reset on reconnect

gatewaySends :: Chan GatewaySendable

Things the library needs to send. Resets to empty on reconnect

startSendingUser :: IORef Bool

If we're really authenticated yet

sendslastStatus :: IORef (Maybe UpdateStatusOpts)

the last sent status

sendlog :: Chan Text

Log

sendableLoop :: Connection -> Sendables -> IO ()