Discord.Internal.Rest

Description

Provides a higher level interface to the rest functions. Preperly writes to the rate-limit loop. Creates separate MVars for each call

Documentation

module Discord.Internal.Types

type DiscordHandleRestChan = Chan (String, JsonRequest, MVar (Either RestCallInternalException ByteString))

class Request a where

Methods

majorRoute :: a -> String

jsonRequest :: a -> JsonRequest

writeRestCall :: (Request (r a), FromJSON a) => DiscordHandleRestChan -> r a -> IO (Either RestCallInternalException a)

Execute a request blocking until a response is received

startRestThread :: Auth -> Chan Text -> IO (DiscordHandleRestChan, ThreadId)

Starts the http request thread. Please only call this once

data RestCallInternalException

Constructors

RestCallInternalErrorCode Int ByteString ByteString

RestCallInternalNoParse String ByteString

RestCallInternalHttpException HttpException