Discord.Internal.Types

Description

Provides types and encoding/decoding code. Types should be identical to those provided in the Discord API documentation.

Documentation

module Discord.Internal.Types.Prelude

module Discord.Internal.Types.Channel

module Discord.Internal.Types.Events

module Discord.Internal.Types.Gateway

module Discord.Internal.Types.Guild

module Discord.Internal.Types.User

module Discord.Internal.Types.Embed

type Object = HashMap Text Value

A JSON "object" (key/value map).

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)