Discord.Internal.Types.Channel

Description

Data structures pertaining to Discord Channels

Documentation

data Channel

Guild channels represent an isolated set of users and messages in a Guild (Server)

Constructors

ChannelText
A text channel in a guild.

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelGuild :: GuildId

The id of the guild.

channelName :: Text

The name of the guild (2 - 1000 characters).

channelPosition :: Integer

The storing position of the channel.

channelPermissions :: [Overwrite]

An array of permission Overwrites

channelUserRateLimit :: Integer

Seconds before a user can speak again

channelNSFW :: Bool

Is not-safe-for-work

channelTopic :: Text

The topic of the channel. (0 - 1024 chars).

channelLastMessage :: Maybe MessageId

The id of the last message sent in the channel

parentId :: Maybe ParentId

The id of the parent channel (category)

ChannelNews

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelGuild :: GuildId

The id of the guild.

channelName :: Text

The name of the guild (2 - 1000 characters).

channelPosition :: Integer

The storing position of the channel.

channelPermissions :: [Overwrite]

An array of permission Overwrites

channelNSFW :: Bool

Is not-safe-for-work

channelTopic :: Text

The topic of the channel. (0 - 1024 chars).

channelLastMessage :: Maybe MessageId

The id of the last message sent in the channel

ChannelStorePage

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelGuild :: GuildId

The id of the guild.

channelName :: Text

The name of the guild (2 - 1000 characters).

channelPosition :: Integer

The storing position of the channel.

channelNSFW :: Bool

Is not-safe-for-work

channelPermissions :: [Overwrite]

An array of permission Overwrites

parentId :: Maybe ParentId

The id of the parent channel (category)

ChannelVoice
A voice channel in a guild.

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelGuild :: GuildId

The id of the guild.

channelName :: Text

The name of the guild (2 - 1000 characters).

channelPosition :: Integer

The storing position of the channel.

channelPermissions :: [Overwrite]

An array of permission Overwrites

channelNSFW :: Bool

Is not-safe-for-work

channelBitRate :: Integer

The bitrate (in bits) of the channel.

channelUserLimit :: Integer

The user limit of the voice channel.

parentId :: Maybe ParentId

The id of the parent channel (category)

ChannelDirectMessage
DM Channels represent a one-to-one conversation between two users, outside the scope of guilds

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelRecipients :: [User]

The User object(s) of the DM recipient(s).

channelLastMessage :: Maybe MessageId

The id of the last message sent in the channel

ChannelGroupDM

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelRecipients :: [User]

The User object(s) of the DM recipient(s).

channelLastMessage :: Maybe MessageId

The id of the last message sent in the channel

ChannelGuildCategory

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelGuild :: GuildId

The id of the guild.

channelName :: Text

The name of the guild (2 - 1000 characters).

channelPosition :: Integer

The storing position of the channel.

channelPermissions :: [Overwrite]

An array of permission Overwrites

ChannelStage

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelGuild :: GuildId

The id of the guild.

channelStageId :: StageId

channelStageTopic :: Text

ChannelUnknownType

Fields

channelId :: ChannelId

The id of the channel (Will be equal to the guild if it's the "general" channel).

channelJSON :: Text

channelIsInGuild :: Channel -> Bool

If the channel is part of a guild (has a guild id field)

data Overwrite

Permission overwrites for a channel.

Constructors

Overwrite

Fields

overwriteId :: OverwriteId

Role or User id

overwriteType :: Text

Either "role" or "member

overwriteAllow :: Integer

Allowed permission bit set

overwriteDeny :: Integer

Denied permission bit set

data Message

Represents information about a message in a Discord channel.

Constructors

Message

Fields

messageId :: MessageId

The id of the message

messageChannel :: ChannelId

Id of the channel the message was sent in

messageAuthor :: User

The User the message was sent by

messageText :: Text

Contents of the message

messageTimestamp :: UTCTime

When the message was sent

messageEdited :: Maybe UTCTime

When/if the message was edited

messageTts :: Bool

Whether this message was a TTS message

messageEveryone :: Bool

Whether this message mentions everyone

messageMentions :: [User]

Users specifically mentioned in the message

messageMentionRoles :: [RoleId]

Roles specifically mentioned in the message

messageAttachments :: [Attachment]

Any attached files

messageEmbeds :: [Embed]

Any embedded content

messageReactions :: [MessageReaction]

Any reactions to message

messageNonce :: Maybe Nonce

Used for validating if a message was sent

messagePinned :: Bool

Whether this message is pinned

messageGuild :: Maybe GuildId

The guild the message went to

messageReference :: Maybe MessageReference

Reference IDs of the original message

referencedMessage :: Maybe Message

The full original message

data MessageReaction

Constructors

MessageReaction

Fields

messageReactionCount :: Int

messageReactionMeIncluded :: Bool

messageReactionEmoji :: Emoji

data Emoji

Represents an emoticon (emoji)

Constructors

Emoji

Fields

emojiId :: Maybe EmojiId

The emoji id

emojiName :: Text

The emoji name

emojiRoles :: Maybe [RoleId]

Roles the emoji is active for

emojiUser :: Maybe User

User that created this emoji

emojiManaged :: Maybe Bool

Whether this emoji is managed

data Attachment

Represents an attached to a message file.

Constructors

Attachment

Fields

attachmentId :: Snowflake

Attachment id

attachmentFilename :: Text

Name of attached file

attachmentSize :: Integer

Size of file (in bytes)

attachmentUrl :: Text

Source of file

attachmentProxy :: Text

Proxied url of file

attachmentHeight :: Maybe Integer

Height of file (if image)

attachmentWidth :: Maybe Integer

Width of file (if image)

newtype Nonce

Constructors

Nonce Text

data MessageReference

Represents a Message Reference

Constructors

MessageReference

Fields

referenceMessageId :: Maybe MessageId

id of the originating message

referenceChannelId :: Maybe ChannelId

id of the originating message's channel

referenceGuildId :: Maybe GuildId

id of the originating message's guild

failIfNotExists :: Bool

Whether to not send if reference not exist