Discord.Internal.Types.Guild

Description

Types relating to Discord Guilds (servers)

Documentation

data GuildMember

Representation of a guild member.

Constructors

GuildMember

Fields

memberUser :: User

memberNick :: Maybe Text

memberRoles :: [Snowflake]

memberJoinedAt :: UTCTime

memberDeaf :: Bool

memberMute :: Bool

data Guild

Guilds in Discord represent a collection of users and channels into an isolated Server

Constructors

Guild

Fields

guildId :: GuildId

Gulid id

guildName :: Text

Guild name (2 - 100 chars)

guildIcon :: Maybe Text

Icon hash

guildSplash :: Maybe Text

Splash hash

guildOwnerId :: UserId

Guild owner id

guildPermissions :: Maybe Integer

guildRegion :: Text

Guild voice region

guildAfkId :: Maybe ChannelId

Id of afk channel

guildAfkTimeout :: Integer

Afk timeout in seconds

guildEmbedEnabled :: Maybe Bool

Id of embedded channel

guildEmbedChannel :: Maybe ChannelId

Id of embedded channel

guildVerificationLevel :: Integer

Level of verification

guildNotification :: Integer

Level of default notifications

guildExplicitFilterLevel :: Integer

guildRoles :: [Role]

Array of Role objects

guildEmojis :: [Emoji]

Array of Emoji objects

guildFeatures :: [Text]

guildMultiFactAuth :: !Integer

guildApplicationId :: Maybe Snowflake

data GuildUnavailable

Constructors

GuildUnavailable

Fields

idOnceAvailable :: GuildId

data GuildInfo

Constructors

GuildInfo

Fields

guildJoinedAt :: UTCTime

guildLarge :: Bool

guildMemberCount :: Integer

guildMembers :: [GuildMember]

guildChannels :: [Channel]

Channels in the guild (sent in GuildCreate) , guildPresences :: [Presence]

data PartialGuild

Constructors

PartialGuild

Fields

partialGuildId :: GuildId

partialGuildName :: Text

partialGuildIcon :: Maybe Text

partialGuildOwner :: Bool

partialGuildPermissions :: Integer

data Role

Roles represent a set of permissions attached to a group of users. Roles have unique names, colors, and can be "pinned" to the side bar, causing their members to be listed separately. Roles are unique per guild, and can have separate permission profiles for the global context (guild) and channel context.

Constructors

Role

Fields

roleId :: RoleId

The role id

roleName :: Text

The role name

roleColor :: Integer

Integer representation of color code

roleHoist :: Bool

If the role is pinned in the user listing

rolePos :: Integer

Position of this role

rolePerms :: Integer

Permission bit set

roleManaged :: Bool

Whether this role is managed by an integration

roleMention :: Bool

Whether this role is mentionable

data VoiceRegion

VoiceRegion is only refrenced in Guild endpoints, will be moved when voice support is added

Constructors

VoiceRegion

Fields

voiceRegionId :: Text

Unique id of the region

voiceRegionName :: Text

Name of the region

voiceRegionVip :: Bool

True if this is a VIP only server

voiceRegionOptimal :: Bool

True for the closest server to a client

voiceRegionDeprecated :: Bool

Whether this is a deprecated region

voiceRegionCustom :: Bool

Whether this is a custom region

data GuildBan

Info about a Ban

Constructors

GuildBan

Fields

guildBanReason :: Text

guildBanUser :: User

data Invite

Represents a code to add a user to a guild

Constructors

Invite

Fields

inviteCode :: Text

The invite code

inviteGuildId :: Maybe GuildId

The guild the code will invite to

inviteChannelId :: ChannelId

The channel the code will invite to

data InviteWithMeta

Invite code with additional metadata

Constructors

InviteWithMeta Invite InviteMeta

data InviteMeta

Additional metadata about an invite.

Constructors

InviteMeta

Fields

inviteCreator :: User

The user that created the invite

inviteUses :: Integer

Number of times the invite has been used

inviteMax :: Integer

Max number of times the invite can be used

inviteAge :: Integer

The duration (in seconds) after which the invite expires

inviteTemp :: Bool

Whether this invite only grants temporary membership

inviteCreated :: UTCTime

When the invite was created

inviteRevoked :: Bool

If the invite is revoked

data Integration

Represents the behavior of a third party account link.

Constructors

Integration

Fields

integrationId :: !Snowflake

Integration id

integrationName :: Text

Integration name

integrationType :: Text

Integration type (Twitch, Youtube, ect.)

integrationEnabled :: Bool

Is the integration enabled

integrationSyncing :: Bool

Is the integration syncing

integrationRole :: RoleId

Id the integration uses for "subscribers"

integrationBehavior :: Integer

The behavior of expiring subscribers

integrationGrace :: Integer

The grace period before expiring subscribers

integrationOwner :: User

The user of the integration

integrationAccount :: IntegrationAccount

The account the integration links to

integrationSync :: UTCTime

When the integration was last synced

data IntegrationAccount

Represents a third party account link.

Constructors

IntegrationAccount

Fields

accountId :: Text

The id of the account.

accountName :: Text

The name of the account.

data GuildEmbed

Represents an image to be used in third party sites to link to a discord channel

Constructors

GuildEmbed

Fields

embedEnabled :: Bool

Whether the embed is enabled

embedChannel :: ChannelId

The embed channel id