Discord.Internal.Types.Prelude

Description

Provides base types and utility functions needed for modules in Discord.Internal.Types

Documentation

data Auth

Authorization token for the Discord API

Constructors

Auth Text

authToken :: Auth -> Text

Get the raw token formatted for use with the websocket gateway

newtype Snowflake

A unique integer identifier. Can be used to calculate the creation date of an entity.

Constructors

Snowflake Word64

type ChannelId = Snowflake

type StageId = Snowflake

type GuildId = Snowflake

type MessageId = Snowflake

type EmojiId = Snowflake

type UserId = Snowflake

type OverwriteId = Snowflake

type RoleId = Snowflake

type IntegrationId = Snowflake

type WebhookId = Snowflake

type ParentId = Snowflake

snowflakeCreationDate :: Snowflake -> UTCTime

Gets a creation date from a snowflake.

epochTime :: UTCTime

Default timestamp